phoenix_test-new/BackEnd/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-dotenv
psycopg2